refactor: Separate GattLiveData from MonitorViewModel as singleton

+++ app/src/main/java/com/seoultech/ecgmonitor/bluetooth/connect/BluetoothConnectStateCallback.kt
@@ -3,6 +3,6 @@ package com.seoultech.ecgmonitor.bluetooth.connect
 interface BluetoothConnectStateCallback {
     fun onConnected()
     fun onDisconnected()
-    fun onValueChanged(value: Int)
+    fun onValueChanged(value: Float)
     fun onFailure()
 }
